Remove API for clearcode and matchcode #241 #445 #453
