+++ docs/source/purldb/api.rst
@@ -916,6 +916,7 @@ Return a list of package sets and the package data of packages within
 ``GET /api/projects/0bbdcf88-ad07-4970-9272-7d5f4c82cc7b/``
 
 .. code-block:: json
+
     {
         "count": 8198,
         "next": "https://public.purldb.io/api/package_sets/?page=2",
@@ -1100,3 +1101,16 @@ Return a ``git_repo`` from a standard PackageURL ``package_url``.
     {
         "git_repo": "git+https://github.com/ckeditor/ckeditor4-react.git"
     }
+
+
+matching
+--------
+
+
+.. code-block:: json
+
+{
+    "upload_file": null,
+    "input_urls": [],
+    "webhook_url": ""
+}
